Changed Datatypes of Ids to integer

+++ LuxuryTravelConcierge/src/main/java/com/revature/DAOS/BookingDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BookingDAO extends JpaRepository<Booking, Long> {
-    List<Booking> findByUserId(Long userId);
-    List<Booking> findByRoomId(Long roomId);
+public interface BookingDAO extends JpaRepository<Booking, Integer> {
+    List<Booking> findByUserId(Integer userId);
+    List<Booking> findByRoomId(Integer roomId);
 }
