+++ LuxuryTravelConcierge/src/main/java/com/revature/models/Booking.java
@@ -10,47 +10,47 @@ public class Booking {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long bookingId;
+    private Integer bookingId;
 
     //Will use a foreign key to room id later
     @OneToOne
     @JoinColumn(name = "roomId", nullable = false)
-    private Long roomId;
+    private Integer roomId;
 
     //Will need to link this to a user
     @ManyToOne
     @JoinColumn(name = "userId", nullable = false)
-    private Long userId;
+    private Integer userId;
 
     //Can use SQL TO_Date to convert varchar to dates
     private String checkInDate;
 
     private String checkOutDate;
 
-    private Long price;
+    private Integer price;
     
 
-    public Long getBookingId() {
+    public Integer getBookingId() {
         return bookingId;
     }
 
-    public void setBookingId(Long bookingId) {
+    public void setBookingId(Integer bookingId) {
         this.bookingId = bookingId;
     }
 
-    public Long getRoomId() {
+    public Integer getRoomId() {
         return roomId;
     }
 
-    public void setRoomId(Long roomId) {
+    public void setRoomId(Integer roomId) {
         this.roomId = roomId;
     }
     
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
 
@@ -71,11 +71,11 @@ public class Booking {
     }
 
 
-    public Long getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
